fix(mdm): address device key review findings
